Update readme and set up project structure
